refactoring background worker
