--- src/scripts/background/index.ts
@@ -1,1 +1,0 @@
-export * from './bgFgRules'
