--- src/scripts/index.ts
@@ -1,1 +1,0 @@
-export * as background from './background';
